+++ R/badge.R
@@ -2,7 +2,10 @@
 #'
 #' These helpers produce the markdown text you need in your README to include
 #' badges that report information, such as the CRAN version or test coverage,
-#' and link out to relevant external resources.
+#' and link out to relevant external resources. To add badges automatically
+#' ensure your badge block starts with a line containing only
+#' `<!-- badges: start -->` and ends with a line containing only
+#' `<!-- badges: end -->`.
 #'
 #' @details
 #'
@@ -41,18 +44,49 @@ NULL
 #' @rdname badges
 #' @export
 use_badge <- function(badge_name, href, src) {
-  if (has_badge(href)) {
+  link <- glue("[![{badge_name}]({src})]({href})")
+  badge_lines <- badge_lines()
+
+  if (is.null(badge_lines)) {
+    if (has_badge(href)) {
+      return(invisible(FALSE))
+    }
+
+    todo(
+      "Add a {field(badge_name)} badge by adding the following line ",
+      "to your README:"
+    )
+    code_block(link)
+    todo(
+      "Allow badges to be added automatically by adding ",
+      "{code(badge_start)} and {code(badge_end)} before and after your badges."
+    )
     return(invisible(FALSE))
   }
 
-  img <- glue("![{badge_name}]({src})")
-  link <- glue("[{img}]({href})")
+  path <- find_readme()
+  lines <- readLines(path)
+  start <- badge_lines[[1]]
+  end <- badge_lines[[2]]
 
-  todo(
-    "Add a {field(badge_name)} badge by adding the following line ",
-    "to your README:"
+  badges <- lines[seq2(start, end)]
+  if (link %in% badges) {
+    return(invisible(FALSE))
+  }
+
+  done("Adding a {field(badge_name)} badge to {value(proj_rel_path(path))}")
+  badges <- c(badges, link)
+  lines <- c(
+    lines[seq2(1, start - 1L)],
+    badges,
+    lines[seq2(end + 1L, length(lines))]
   )
-  code_block(link)
+  writeLines(lines, path)
+
+  if (path_ext(path) == "Rmd") {
+    todo("Re-knit {value(proj_rel_path(path))}")
+  }
+  invisible(TRUE)
 }
 
 #' @rdname badges
@@ -74,12 +108,8 @@ use_bioc_badge <- function() {
   check_is_package("use_bioc_badge()")
   pkg <- project_name()
 
-  src <- glue(
-    "http://www.bioconductor.org/shields/build/release/bioc/{pkg}.svg"
-  )
-  href <- glue(
-    "https://bioconductor.org/checkResults/release/bioc-LATEST/{pkg}"
-  )
+  src <- glue("http://www.bioconductor.org/shields/build/release/bioc/{pkg}.svg")
+  href <- glue("https://bioconductor.org/checkResults/release/bioc-LATEST/{pkg}")
   use_badge("BioC status", href, src)
 
   invisible(TRUE)
@@ -124,8 +154,6 @@ use_binder_badge <- function() {
   invisible(TRUE)
 }
 
-
-
 has_badge <- function(href) {
   readme_path <- proj_path("README.md")
   if (!file_exists(readme_path)) {
@@ -134,4 +162,44 @@ has_badge <- function(href) {
 
   readme <- readLines(readme_path, encoding = "UTF-8")
   any(grepl(href, readme, fixed = TRUE))
+}
+
+# Badge data structure ----------------------------------------------------
+
+badge_start <- "<!-- badges: start -->"
+badge_end <- "<!-- badges: end -->"
+
+badge_lines <- function() {
+  path <- find_readme()
+  if (is.null(path))
+    return(NULL)
+
+  lines <- readLines(path)
+  start <- which(lines == badge_start)
+  end <- which(lines == badge_end)
+
+  # No badges
+  if (length(start) == 0 && length(end) == 0)
+    return(NULL)
+
+  if (!(length(start) == 1 && length(end) == 1 && start < end)) {
+    stop_glue(
+      "Invalid badge specification.
+      Must start with {code(badge_start)} and end with {code(badge_end)}"
+    )
+  }
+
+  c(start + 1L, end - 1L)
+}
+
+find_readme <- function() {
+  Rmd <- proj_path("README.Rmd")
+  if (file_exists(Rmd))
+    return(Rmd)
+
+  md <- proj_path("README.md")
+  if (file_exists(md))
+    return(md)
+
+  NULL
 }